I forgot what i changed. Fixed mass loss to follow CBC durabilityPenalty formula and rebalanced autocannon penetration, same as main

+++ src/main/java/com/cbc_terminal_ballistics/state/ImpactMark.java
@@ -34,6 +34,10 @@ public record ImpactMark(ImpactMarkKind kind, TBCaliber caliber, ImpactSurfaceTy
     }
 
     public Vec3 absolute(net.minecraft.core.BlockPos pos) {
-        return new Vec3(pos.getX() + x, pos.getY() + y, pos.getZ() + z);
+        // Cast to double before adding so we don't lose precision
+        // at massive Valkyrien Skies shipyard coordinates!
+        return new Vec3(pos.getX() + (double) x,
+                pos.getY() + (double) y,
+                pos.getZ() + (double) z);
     }
 }
